+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+- Add safety check to mutation observer
+
 ### Upgrades
 
 - Upgrade JUnit to 5.9.2